Add operator js to components director
